requirements.txt
# ── Django Core ──────────────────────────────
django==5.0.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
dj-database-url==2.2.0
python-decouple==3.8

# ── Auth ─────────────────────────────────────
djangorestframework-simplejwt==5.3.1

# ── Caching ──────────────────────────────────
django-redis==5.4.0
redis==5.0.7

# ── Async Tasks ──────────────────────────────
celery==5.4.0
django-celery-beat==2.6.0

# ── Kafka ────────────────────────────────────
confluent-kafka==2.4.0

# ── Payments ─────────────────────────────────
stripe==10.5.0

# ── AI Agent ─────────────────────────────────
langchain==0.2.11
langchain-google-genai==1.0.7
httpx==0.27.0

# ── Filters & Utils ──────────────────────────
django-filter==24.2
Pillow==10.4.0

# ── Production Server ─────────────────────────
gunicorn==22.0.0

# ── Dev & Testing ────────────────────────────
pytest-django==4.8.0
pytest==8.3.1
factory-boy==3.3.1
faker==26.0.0
coverage==7.6.0
django-cors-headers==4.3.1